+++ tests/DecimalNumberTest.php
@@ -13,22 +13,22 @@ declare(strict_types=1);
 
 namespace BakameTest\Period\Visualizer;
 
-use Bakame\Period\Visualizer\IntegerLabel;
+use Bakame\Period\Visualizer\DecimalNumber;
 use League\Period\Period;
 use League\Period\Sequence;
 use PHPUnit\Framework\TestCase;
 
 /**
- * @coversDefaultClass \Bakame\Period\Visualizer\IntegerLabel;
+ * @coversDefaultClass \Bakame\Period\Visualizer\DecimalNumber;
  */
-final class IntegerGeneratorTest extends TestCase
+final class DecimalNumberTest extends TestCase
 {
     /**
      * @dataProvider providerLetter
      */
     public function testGetLabels(Sequence $sequence, int $label, array $expected): void
     {
-        $generator = new IntegerLabel($label);
+        $generator = new DecimalNumber($label);
         self::assertSame($expected, $generator->generate($sequence));
     }
 
@@ -68,19 +68,48 @@ final class IntegerGeneratorTest extends TestCase
 
     public function testStartWith(): void
     {
-        $generator = new IntegerLabel(42);
+        $generator = new DecimalNumber(42);
         self::assertSame(42, $generator->startingAt());
         $new = $generator->startsWith(69);
         self::assertNotSame($new, $generator);
         self::assertSame(69, $new->startingAt());
         self::assertSame($generator, $generator->startsWith(42));
-        self::assertSame(1, (new IntegerLabel(-3))->startingAt());
+        self::assertSame(1, (new DecimalNumber(-3))->startingAt());
         self::assertSame(1, $generator->startsWith(-3)->startingAt());
     }
 
     public function testFormat(): void
     {
-        $generator = new IntegerLabel(42);
+        $generator = new DecimalNumber(42);
         self::assertSame('', $generator->format([]));
+    }
+
+    public function testZeroLeftPaddingSetterAndStatus(): void
+    {
+        $generator = new DecimalNumber(42, 42);
+        self::assertFalse($generator->isPadded());
+
+        $newGenerator = $generator->withPadding();
+        self::assertTrue($newGenerator->isPadded());
+        self::assertSame($newGenerator, $newGenerator->withPadding());
+
+        $wrongGenerator = $newGenerator->withoutPadding();
+        self::assertFalse($wrongGenerator->isPadded());
+        self::assertSame($wrongGenerator, $wrongGenerator->withoutPadding());
+    }
+
+    public function testAddingLeadingZero(): void
+    {
+        $generator = new DecimalNumber(98, DecimalNumber::LEFT_PAD);
+        $labels = $generator->generate(new Sequence(
+            Period::before('yesterday', '1 DAY'),
+            Period::after('yesterday', '1 DAY'),
+            Period::around('yesterday', '1 DAY')
+        ));
+
+        /** @var string $label */
+        foreach ($labels as $label) {
+            self::assertSame(3, strlen($label));
+        }
     }
 }